auth config

+++ AuthService/src/main/java/com/easyrun/auth/StartupInit.java
@@ -4,6 +4,7 @@ import javax.annotation.PostConstruct;
 
 import org.jose4j.jwk.RsaJsonWebKey;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import com.easyrun.commons.model.Configuration;
@@ -17,15 +18,19 @@ public class StartupInit {
 	@Autowired
 	private RsaJsonWebKey rsaJsonWebKey;
 	
-	
-	
+	@Value("${isMaster}")
+	private String isMaster;
+		
 	@PostConstruct
 	public void init() throws Exception {
-		configurationRepository.deleteAll();
-		Configuration c = new Configuration();
-		c.setPublicKey(rsaJsonWebKey.toJson());
-		c.setHeaderKeyId(rsaJsonWebKey.getKeyId());
-		c.setName("PUBLIC_KEY");
-		configurationRepository.insert(c);
+		boolean isMasterFlag = Boolean.valueOf(isMaster);
+		if (isMasterFlag) {
+			configurationRepository.deleteAll();
+			Configuration c = new Configuration();			
+			c.setPublicKey(rsaJsonWebKey.toJson());
+			c.setHeaderKeyId(rsaJsonWebKey.getKeyId());
+			c.setName("PUBLIC_KEY");
+			configurationRepository.insert(c);
+		}
 	}
 }
